fix(src): Riscrittura moduli
Eliminata cartella inutile e sostituita con file python
