--- cnlp/utils/__init__.py
@@ -1,4 +1,0 @@
-"""Some utilities functions."""
-from cnlp.utils.nodes_to_indexes import nodes_to_indexes
-from cnlp.utils.to_adjacency_matrix import to_adjacency_matrix
-from cnlp.utils.only_unconnected import only_unconnected
