+++ docs/index.html
@@ -53,7 +53,7 @@ import cnlp.utils</code></pre>
 <dd>
 <div class="desc"><p>Similarity based Methods for Link Prediction …</p></div>
 </dd>
-<dt><code class="name"><a title="cnlp.utils" href="utils/index.html">cnlp.utils</a></code></dt>
+<dt><code class="name"><a title="cnlp.utils" href="utils.html">cnlp.utils</a></code></dt>
 <dd>
 <div class="desc"><p>Some utilities functions.</p></div>
 </dd>
@@ -78,7 +78,7 @@ import cnlp.utils</code></pre>
 <li><code><a title="cnlp.other_methods" href="other_methods/index.html">cnlp.other_methods</a></code></li>
 <li><code><a title="cnlp.probabilistic_methods" href="probabilistic_methods/index.html">cnlp.probabilistic_methods</a></code></li>
 <li><code><a title="cnlp.similarity_methods" href="similarity_methods/index.html">cnlp.similarity_methods</a></code></li>
-<li><code><a title="cnlp.utils" href="utils/index.html">cnlp.utils</a></code></li>
+<li><code><a title="cnlp.utils" href="utils.html">cnlp.utils</a></code></li>
 </ul>
 </li>
 </ul>
